add multiple files
