+++ CoinCollect/functions.cpp
@@ -1,0 +1,22 @@
+#include<chrono>
+#include<random>
+#include"function.hpp"
+
+unsigned int genCoinRow(){
+  std::random_device rd{};
+  std::seed_seq ss{static_cast<unsigned int>(std::chrono::steady_clock::now().time_since_epoch().count()),rd(),rd(),rd(),rd(),rd(),rd(),rd()};
+  std::mt19937 mt{ss};
+  std::uniform_int_distribution genPos{0,19};
+  return genPos(mt);
+}
+
+
+unsigned int genCoinCol(){
+  std::random_device rd{};
+  std::seed_seq ss{static_cast<unsigned int>(std::chrono::steady_clock::now().time_since_epoch().count()),rd(),rd(),rd(),rd(),rd(),rd(),rd()};
+  std::mt19937 mt{ss};
+  std::uniform_int_distribution genPos{0,49};
+  return genPos(mt);
+}
+
+
